Refactor: move reading statistics and history listing out of ModelData
- Migrated getReadingStatistics to a static method in BookDeviceReadingPositionHistory extension.
- Moved listBookDeviceReadingPositionHistory logic to ReadingSessionManager.
- Updated ModelData to act as a bridge for these methods to maintain backward compatibility.
- Added RealmSwift import to ReadingSessionManager.
